requirements.txt
Flask==2.3.2
Flask_Cors==4.0.0
matplotlib==3.7.2
mysql_connector_repackaged==0.3.1
networkx==3.1
openai==0.27.8
pandas==2.0.3
PyPDF2==3.0.1
scikit_learn==1.3.0
seaborn==0.12.2
